requirements.txt
arabic-reshaper==3.0.0
cryptg==0.5.2
pyaes==1.6.1
pyasn1==0.6.3
python-bidi==0.6.7
python-dotenv==1.2.2
rsa==4.9.1
Telethon==1.42.0
